+++ addons/confirma/src/extensions/ConfirmBooleanExtensions.cs
@@ -6,15 +6,33 @@ public static class ConfirmBooleanExtensions
 {
     public static bool ConfirmTrue(this bool actual, string? message = null)
     {
-        return actual
-            ? actual
-            : throw new ConfirmAssertException(message ?? "Expected true but was false.");
+        if (actual)
+        {
+            return actual;
+        }
+
+        return message is not null
+            ? throw new ConfirmAssertException(message)
+            : throw new ConfirmAssertException(
+                nameof(ConfirmTrue),
+                true,
+                actual
+            );
     }
 
     public static bool ConfirmFalse(this bool actual, string? message = null)
     {
-        return !actual
-            ? actual
-            : throw new ConfirmAssertException(message ?? "Expected false but was true.");
+        if (!actual)
+        {
+            return actual;
+        }
+
+        return message is not null
+            ? throw new ConfirmAssertException(message)
+            : throw new ConfirmAssertException(
+                nameof(ConfirmFalse),
+                false,
+                actual
+            );
     }
 }