+++ CHANGELOG.md
@@ -11,7 +11,6 @@ All notable changes to this project will be documented in this file.
 - Warning when trying to run tests without '--confirma-run' argument.
 - `WhenNotRunningCategory` value to `EIgnoreMode` enum.
 - `HideFromResults` field to `Ignore` attribute.
-- CollectionHelper class.
 - AssertionMessageGenerator class.
 - IFormatter interface.
 - Formatters:
@@ -19,6 +18,8 @@ All notable changes to this project will be documented in this file.
   - NumericFormatter
   - StringFormatter
   - CollectionFormatter
+- CollectionHelper class.
+- Constructor to ConfirmAssertException which automatically generates a message.
 
 ### Changed